src/BinaryKits.Zpl.Protocol.UnitTest/Commands/FieldSeparatorCommandTest.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;

namespace BinaryKits.Zpl.Protocol.Commands.UnitTest
{
    [TestClass]
    public class FieldSeparatorCommandTest
    {
        [TestMethod]
        public void ToZpl_Default_Successful()
        {
            var command = new FieldSeparatorCommand();
            var zplCommand = command.ToZpl();
            Assert.AreEqual("^FS", zplCommand);
        }

        [TestMethod]
        public void IsCommandParsable_ValidCommand_True()
        {
            var isParsable = FieldSeparatorCommand.CanParseCommand("^FS");
            Assert.IsTrue(isParsable);
        }

        [TestMethod]
        public void IsCommandParsable_InvalidCommand_False()
        {
            var isParsable = FieldSeparatorCommand.CanParseCommand("^FT10,10");
            Assert.IsFalse(isParsable);
        }

        [TestMethod]
        public void ParseCommand_ValidCommand1_Successful()
        {
            var command = CommandBase.ParseCommand("^FS");
            Assert.IsTrue(command is FieldSeparatorCommand);
        }

    }
}
